Move sqldialect to types module
